Added a Mapsui project for drawing

+++ VectorTiles/EvaluationContext.cs
@@ -7,11 +7,11 @@
     {
         public float? Zoom { get; set; }
 
-        public int Scale { get; set; }
+        public float Scale { get; set; }
 
         public TagsCollection Tags { get; set; }
 
-        public EvaluationContext(float? zoom, int scale = 1, TagsCollection tags = null)
+        public EvaluationContext(float? zoom, float scale = 1, TagsCollection tags = null)
         {
             Zoom = zoom;
             Scale = scale;
